+++ src/main/java/com/kodea/tdd/calc/oper/Operador.java
@@ -4,19 +4,19 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class Operador  implements IOperador{
-	public double add(int val1, int val2) {
+	public double add(double val1, double val2) {
 		return val1+val2;
 	}
 
-	public double subtract(int val1, int val2) {
+	public double subtract(double val1, double val2) {
 		return val1-val2;
 	}
 
-	public double multiply(int val1, int val2) {
+	public double multiply(double val1, double val2) {
 		return val1*val2;
 	}
 
-	public double divide(int val1, int val2) throws ArithmeticException{
+	public double divide(double val1, double val2) throws ArithmeticException{
 		if (val2 == 0) {
 			throw new ArithmeticException("Segundo operando debe ser distinto de cero.");
 		}
